DLL Initialization Thread
The DLL_THREAD_ATTACH and DLL_THREAD_DETACH cases have been added to the switch statement. These cases are called when a new thread is created or destroyed within the process. Since this code doesn't need to do anything in those cases, they simply break out of the switch statement.